+++ Cargo.toml
@@ -1,13 +1,14 @@
-[package]
-name = "length-log"
-version = "0.1.0"
-authors = ["Kristoffer Andersson <kod.kristoff@gmail.com>"]
-edition = "2021"
+[workspace]
+resolver = "2"
+members = ["crates/*"]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[workspace.dependencies]
+# local crates
+length-log-core = { path = "./crates/length-log-core" }
+length-log-repl = { path = "./crates/length-log-repl" }
 
-[dependencies]
-polars = { version = "0.38", default-features = false }
+# external crates
+# polars = { version = "0.38", default-features = false }
 rustyline = "10"
 # log = "0"
 # env_logger = "0"
