+++ crates/length-log-core-impl/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "length-log-core-impl"
+version = "0.1.0"
+authors = ["Kristoffer Andersson <kod.kristoff@gmail.com>"]
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log.workspace = true
+# polars.workspace = true
