+++ crates/length-log-core/src/lib.rs
@@ -1,0 +1,3 @@
+mod app;
+
+pub use app::App;
