+++ crates/length-log-repl/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "length-log-repl"
+version = "0.1.0"
+authors = ["Kristoffer Andersson <kod.kristoff@gmail.com>"]
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rustyline.workspace = true
+# nom = "7.1"
+shlex.workspace = true
+log.workspace = true
+xflags.workspace = true
+
+length-log-core.workspace = true
