+++ crates/length-log-repl/src/lib.rs
@@ -1,0 +1,7 @@
+// pub mod app;
+pub mod error;
+pub mod repl;
+
+// pub use app::App;
+pub use error::Error;
+pub use repl::run_repl;
