+++ crates/length-log/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "length-log"
+version = "0.1.0"
+authors = ["Kristoffer Andersson <kod.kristoff@gmail.com>"]
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log.workspace = true
+env_logger.workspace = true
+
+length-log-core.workspace = true
+length-log-repl.workspace = true
